rustdoc_json_types
Enum ItemEnum

pub enum ItemEnum {

    Module(Module),
    ExternCrate {
        name: String,
        rename: Option<String>,
    },
    Use(Use),
    Union(Union),
    Struct(Struct),
    StructField(Type),
    Enum(Enum),
    Variant(Variant),
    Function(Function),
    Trait(Trait),
    TraitAlias(TraitAlias),
    Impl(Impl),
    TypeAlias(TypeAlias),
    Constant {
        type_: Type,
        const_: Constant,
    },
    Static(Static),
    ExternType,
    Macro(String),
    ProcMacro(ProcMacro),
    Primitive(Primitive),
    AssocConst {
        type_: Type,
        value: Option<String>,
    },
    AssocType {
        generics: Generics,
        bounds: Vec<GenericBound>,
        type_: Option<Type>,
    },

}

Specific fields of an item.

Part of Item.

Variants

Module(Module)

A module declaration, e.g. mod foo; or mod foo {}

ExternCrate

A crate imported via the extern crate syntax.

Fields

name: String

The name of the imported crate.

rename: Option<String>

If the crate is renamed, this is its name in the crate.

Use(Use)

An import of 1 or more items into scope, using the use keyword.

Union(Union)

A union declaration.

Struct(Struct)

A struct declaration.

StructField(Type)

A field of a struct.

Enum(Enum)

An enum declaration.

Variant(Variant)

A variant of a enum.

Function(Function)

A function declaration (including methods and other associated functions)

Trait(Trait)

A trait declaration.

TraitAlias(TraitAlias)

A trait alias declaration, e.g. trait Int = Add + Sub + Mul + Div;

See the tracking issue

Impl(Impl)

An impl block.

TypeAlias(TypeAlias)

A type alias declaration, e.g. type Pig = std::borrow::Cow<'static, str>;

Constant

The declaration of a constant, e.g. const GREETING: &str = "Hi :3";

Fields

type_: Type

The type of the constant.

const_: Constant

The declared constant itself.

Static(Static)

A declaration of a static.

ExternType

types from an extern block.

See the tracking issue

Macro(String)

A macro_rules! declarative macro. Contains a single string with the source representation of the macro with the patterns stripped.

ProcMacro(ProcMacro)

A procedural macro.

Primitive(Primitive)

A primitive type, e.g. u32.

Items of this kind only come from the core library.

AssocConst

An associated constant of a trait or a type.

Fields

type_: Type

The type of the constant.

value: Option<String>

Inside a trait declaration, this is the default value for the associated constant, if provided. Inside an impl block, this is the value assigned to the associated constant, and will always be present.

The representation is implementation-defined and not guaranteed to be representative of either the resulting value or of the source code.

const X: usize = 640 * 1024;
//               ^^^^^^^^^^

AssocType

An associated type of a trait or a type.

Fields

generics: Generics

The generic parameters and where clauses on ahis associated type.

bounds: Vec<GenericBound>

The bounds for this associated type. e.g.

trait IntoIterator {
    type Item;
    type IntoIter: Iterator<Item = Self::Item>;
//                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
}

type_: Option<Type>

Inside a trait declaration, this is the default for the associated type, if provided. Inside an impl block, this is the type assigned to the associated type, and will always be present.

type X = usize;
//       ^^^^^