rustdoc_json_types
Struct Crate

pub struct Crate {
    pub root: Id,
    pub crate_version: Option<String>,
    pub includes_private: bool,
    pub index: FxHashMap<Id, Item>,
    pub paths: FxHashMap<Id, ItemSummary>,
    pub external_crates: FxHashMap<u32, ExternalCrate>,
    pub format_version: u32,
}

The root of the emitted JSON blob.

It contains all type/documentation information about the language items in the local crate, as well as info about external items to allow tools to find or link to them.

Fields
root: Id

The id of the root Module item of the local crate.

crate_version: Option<String>

The version string given to --crate-version, if any.

includes_private: bool

Whether or not the output includes private items.

index: FxHashMap<Id, Item>

A collection of all items in the local crate as well as some external traits and their items that are referenced locally.

paths: FxHashMap<Id, ItemSummary>

Maps IDs to fully qualified paths and other info helpful for generating links.

external_crates: FxHashMap<u32, ExternalCrate>

Maps crate_id of items to a crate name and html_root_url if it exists.

format_version: u32

A single version number to be used in the future when making backwards incompatible changes to the JSON output.