Crate rustdoc_json_types

Rustdoc’s JSON output interface

These types are the public API exposed through the --output-format json flag. The Crate struct is the root of the JSON blob and all other items are contained within.

Structs

Constant

Crate

A Crate is the root of the emitted JSON blob. It contains all type/documentation information about the language items in the local crate, as well as info about external items to allow tools to find or link to them.

Deprecation

Discriminant

DynTrait

Enum

ExternalCrate

FnDecl

Function

Represents a function (including methods and other associated functions)

FunctionPointer

GenericParamDef

Generics

Header

Id

Impl

Import

Item

ItemSummary

For external (not defined in the local crate) items, you don’t get the same level of information. This struct should contain enough to generate a link/reference to the item in question, or can be used by a tool that takes the json output of multiple crates to find the actual item definition with all the relevant info.

Module

OpaqueTy

Path

PolyTrait

A trait and potential HRTBs

Primitive

ProcMacro

Span

Static

Struct

Trait

TraitAlias

TypeAlias

TypeBinding

Union

Variant

Enums

Abi

GenericArg

GenericArgs

GenericBound

GenericParamDefKind

ItemEnum

ItemKind

MacroKind

StructKind

Term

TraitBoundModifier

Type

TypeBindingKind

VariantKind

Visibility

WherePredicate

Constants

FORMAT_VERSION

rustdoc format-version.